Support for link and update --all as well as resolve tests for postinstall hook

+++ test/helpers.js
@@ -98,9 +98,26 @@ exports.TempDir = (function() {
         return this;
     };
 
-    TempDir.prototype.prepare = function (files) {
-        rimraf.sync(this.path);
-        mkdirp.sync(this.path);
+    TempDir.prototype.delete = function (files) {
+        var that = this;
+
+        if (files) {
+            if (typeof files === 'string') {
+                files = [files];
+            }
+            files.forEach(function(file) {
+                var fullPath = path.join(that.path, file);
+                rimraf.sync(fullPath);
+            });
+        }
+        return this;
+    };
+
+    TempDir.prototype.prepare = function (files, keepFiles) {
+        if (!keepFiles) {
+            rimraf.sync(this.path);
+            mkdirp.sync(this.path);
+        }
         this.create(files);
 
         return this;
